+++ RandomSkunk.Results/Result{T}.cs
@@ -24,11 +24,7 @@ public readonly partial struct Result<T> : IResult<T>, IEquatable<Result<T>>
         _value = default;
         _error = error ?? new Error();
 
-        if (_error.StackTrace is null && !(omitStackTrace ?? FailResult.OmitStackTrace))
-            _error = _error with { StackTrace = FilteredStackTrace.Create() };
-
-        _error = FailResult.InvokeReplaceErrorIfSet(_error);
-        FailResult.InvokeCallbackIfSet(_error);
+        FailResult.Handle(ref _error, omitStackTrace);
     }
 
     private enum Outcome